+++ test/lib/parse/query/constraints/not_in_query_test.rb
@@ -1,0 +1,15 @@
+require_relative '../../../../test_helper'
+
+class TestNotInQueryConstraint < Minitest::Test
+  extend Minitest::Spec::DSL
+  include ConstraintTests
+
+  def setup
+    @klass = Parse::NotInQueryConstraint
+    @key = :$notInQuery
+    @operand = :excludes
+    @keys = [:excludes, :not_in_query]
+    @skip_scalar_values_test = true
+  end
+
+end